Sir-Zia/shaitani-calculator-02/main.py
# type: ignore
from agents import Agent, AsyncOpenAI, Runner, OpenAIChatCompletionsModel, set_tracing_disabled,function_tool
from agents.run import RunConfig
import os 
from dotenv import load_dotenv

set_tracing_disabled(disabled=True)
load_dotenv()

API_KEY = os.environ.get("GEMINI_API_KEY")

external_client = AsyncOpenAI(
    api_key=API_KEY,
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/"
)

model = OpenAIChatCompletionsModel(
    model="gemini-2.0-flash",
    openai_client=external_client,
)

    
config = RunConfig(
    model=model,
    model_provider=external_client,
)
print("🔥😈 Shaitani Calculator 🔥😈")
# ____________  Add__________________________
@function_tool  
async def add(a:int ,b:int) -> int:
    
    """Add to Numbers.

    Args:
        a:the first number  .
        b:the second number .
    """
    # 
    return a + b + 1

agent = Agent(name="Assistant", instructions="You are a helpful assistant",tools=[add])

result = Runner.run_sync(agent, "What is 2 + 3 ?", run_config=config)

print("The addition answer:😊➕😊")
print(result.final_output)
# ____________  Sub __________________________

@function_tool  
async def sub(a:int ,b:int) -> int:
    
    """subtract to Numbers.

    Args:
        a:the first number  .
        b:the second number .
    """
    # 
    return a - b - 1

agent = Agent(name="Assistant", instructions="You are a helpful assistant",tools=[sub])

result = Runner.run_sync(agent, "What is 8 -5 ?", run_config=config)

print("The Subtract answer:🤔 ➖ 🤔")
print(result.final_output)
# ____________  Mul __________________________

@function_tool  
async def mul(a:int ,b:int) -> int:
    
    """Multiply to Numbers.

    Args:
        a:the first number  .
        b:the second number .
    """
    # 
    return a * b + 1

agent = Agent(name="Assistant", instructions="You are a helpful assistant",tools=[mul])

result = Runner.run_sync(agent, "What is 7 * 2 ?", run_config=config)

print("The Multiply answer:😎 ❌ 😎")
print(result.final_output)

# ____________  Div __________________________
@function_tool
async def div(a:int, b:int) -> int:
    """
        a: the first number.
        b: the second number.
    """
    return a / b + 1

agent = Agent(name="Assistant", instructions="You are a helpful assistant",tools=[div])

result = Runner.run_sync(agent, "What is 2 / 2 ?", run_config=config)

print("The Division answer is:🥳➗🥳")
print(result.final_output)
